Cockpit: Moved upgrade runners to their own module.
